ort

Module environment

The Environment is a process-global configuration under which Sessions are created.

With it, you can configure default execution providers, enable/disable telemetry, share a global thread pool across all sessions, or add a custom logger.

Environments can be set up via ort::init:

ort::init()
	.with_execution_providers([
		#[cfg(feature = "cuda")]
		ep::CUDA::default().build()
	])
	.commit();

// ... do other ort things now that our environment is set up...

With the load-dynamic feature, you can also load the runtime from a direct path to a DLL with ort::init_from:

let lib_path = std::env::current_exe().unwrap().parent().unwrap().join("lib");
ort::init_from(lib_path.join("onnxruntime.dll"))?
	.with_execution_providers([
		#[cfg(feature = "cuda")]
		ep::CUDA::default().build()
	])
	.commit();

If you don’t configure an environment, one will be created with default settings at the first creation of a session. The environment can’t be re-configured after one is committed, so it’s important ort::init comes before any other ort API for the config to take effect. Authors of libraries using ort should never have the library configure the environment itself; allow the application developer to do that themselves if they wish.

Structs

Environment

Holds shared global configuration for all Sessions in the process.

EnvironmentBuilder

Struct used to build an Environment; see crate::init.

GlobalThreadPoolOptions

Traits

ThreadManager

Used for customizing the thread spawning process of a global thread pool or session thread pool. Could be used to add additional initialization/cleanup code to inference threads for better debugging/error handling.

Functions

init

Creates an ONNX Runtime environment.

init_from

Creates an ONNX Runtime environment, dynamically loading ONNX Runtime from the library file (.dll/.so/.dylib) specified by path. Returns an error if the dylib fails to load.