ort::environment

Function init_from

pub fn init_from<P: AsRef<Path>>(path: P) -> Result<EnvironmentBuilder>

Creates an ONNX Runtime environment, dynamically loading ONNX Runtime from the library file (.dll/.so/.dylib) specified by path. Returns an error if the dylib fails to load.

This must be called before any other ort APIs are used in order for the correct dynamic library to be loaded.

let lib_path = std::env::current_exe().unwrap().parent().unwrap().join("lib");
ort::init_from(lib_path.join("onnxruntime.dll"))?
	.with_execution_providers([ep::CUDA::default().build()])
	.commit();

Notes

In order for environment settings to apply, this must be called before you use other APIs like Session, and you must call .commit() on the builder returned by this function.