ort

Module ep

ExecutionProviders provide hardware acceleration to Sessions.

Sessions can be configured with execution providers via SessionBuilder::with_execution_providers:

use ort::{ep, session::Session};

fn main() -> ort::Result<()> {
	let session = Session::builder()?
		.with_execution_providers([
			#[cfg(feature = "cuda")]
			ep::CUDA::default().build()
		])?
		.commit_from_file("model.onnx")?;

	Ok(())
}

Re-exports

pub use self::cpu::CPU;

Modules

cpu

Structs

ExecutionProviderDispatch

Dynamic execution provider container, used to provide a list of multiple types of execution providers when configuring execution providers for a SessionBuilder or EnvironmentBuilder.

ExecutionProviderLibrary

Handle to a loaded execution provider library, obtained from Environment::register_ep_library.

Enums

ArenaExtendStrategy

The strategy for extending the device memory arena.

Traits

ArbitrarilyConfigurableExecutionProvider

Trait used for execution providers that can have arbitrary configuration keys applied.

ExecutionProvider

ExecutionProviderResource

Functions

get_gpu_device

Returns the ID of the GPU device being used by the active EP.

set_gpu_device

Sets the current GPU device of the active EP to the device specified by device_id.