ort::ep::cuda

Function preload_dylibs

pub fn preload_dylibs(
    cuda_root_dir: Option<&Path>,
    cudnn_root_dir: Option<&Path>,
) -> Result<()>

Preload the dylibs required by CUDA/cuDNN.

This attempts to load all dynamic libraries required by the CUDA execution provider from the given CUDA and cuDNN directories, if they are provided. Passing None will prevent preloading binaries for that component. This function will immediately return with an error when a library fails to load, without attempting to load the rest of the libraries.

Preloading a library in this way will prioritize it in the search order when the CUDA EP attempts to load its dependencies, effectively allowing you to customize the CUDA install path without modifying the PATH environment variable. Note that this function intentionally leaks memory; see crate::util::preload_dylib for more information.

use ort::ep;

let cuda_root = Path::new(r#"C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v12.6\bin"#);
let cudnn_root = Path::new(r#"D:\cudnn_9.8.0"#);

// Load CUDA & cuDNN
let _ = ep::cuda::preload_dylibs(Some(cuda_root), Some(cudnn_root));

// Only preload cuDNN
let _ = ep::cuda::preload_dylibs(None, Some(cudnn_root));