ort::ep

Trait ArbitrarilyConfigurableExecutionProvider

pub trait ArbitrarilyConfigurableExecutionProvider {
    // Required method
    fn with_arbitrary_config(
        self,
        key: impl ToString,
        value: impl ToString,
    ) -> Self;
}

Trait used for execution providers that can have arbitrary configuration keys applied.

Most execution providers have a small set of configuration options which don’t change between ONNX Runtime releases; others, like the CUDA execution provider, often have options added that go undocumented and thus unimplemented by ort. This allows you to configure these options regardless.

Required Methods

fn with_arbitrary_config(self, key: impl ToString, value: impl ToString) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.