ort::ep

Trait ExecutionProvider

pub trait ExecutionProvider:
    Any
    + Send
    + Sync {
    // Required methods
    fn name(&self) -> &'static str;
    fn register(
        &self,
        session_builder: &mut SessionBuilder,
    ) -> Result<(), RegisterError>;

    // Provided methods
    fn supported_by_platform(&self) -> bool { ... }
    fn is_available(&self) -> Result<bool> { ... }
}

Required Methods

fn name(&self) -> &'static str

Returns the identifier of this execution provider used internally by ONNX Runtime.

This is the same as what’s used in ONNX Runtime’s Python API to register this execution provider, i.e. TVM’s identifier is TvmExecutionProvider.

fn register( &self, session_builder: &mut SessionBuilder, ) -> Result<(), RegisterError>

Attempts to register this execution provider on the given session.

Provided Methods

fn supported_by_platform(&self) -> bool

Returns whether this execution provider is supported on this platform.

For example, the CoreML execution provider implements this as:

impl ExecutionProvider for CoreML {
	fn supported_by_platform() -> bool {
		cfg!(target_vendor = "apple")
	}
}

fn is_available(&self) -> Result<bool>

Returns Ok(true) if ONNX Runtime was compiled with support for this execution provider, and Ok(false) otherwise.

An Err may be returned if a serious internal error occurs, in which case your application should probably just abort.

Note that this does not always mean the execution provider is usable for a specific session. A model may use operators not supported by an execution provider, or the EP may encounter an error while attempting to load dependencies during session creation. In most cases (i.e. showing the user an error message if CUDA could not be enabled), you’ll instead want to manually register this EP via ExecutionProvider::register and detect and handle any errors returned by that function.