ort::ep::tvm

Struct TVM

pub struct TVM {

    pub executor: Option<ExecutorType>,
    pub so_folder: Option<String>,
    pub check_hash: Option<bool>,
    pub hash_file_path: Option<String>,
    pub target: Option<String>,
    pub target_host: Option<String>,
    pub opt_level: Option<usize>,
    pub freeze_weights: Option<bool>,
    pub to_nhwc: Option<bool>,
    pub tuning_type: Option<TuningType>,
    pub tuning_file_path: Option<String>,
    pub input_names: Option<String>,
    pub input_shapes: Option<String>,

}

Fields
executor: Option<ExecutorType>

Executor type used by TVM. There is a choice between two types, GraphExecutor and VirtualMachine. Default is ExecutorType::VirtualMachine.

so_folder: Option<String>

Path to folder with set of files (.ro-, .so/.dll-files and weights) obtained after model tuning.

check_hash: Option<bool>

Whether or not to perform a hash check on the model obtained in the so_folder.

hash_file_path: Option<String>

A path to a file that contains the pre-computed hash for the ONNX model located in the so_folder for checking when check_hash is Some(true).

target: Option<String>target_host: Option<String>opt_level: Option<usize>freeze_weights: Option<bool>

Whether or not all model weights are kept on compilation stage, otherwise they are downloaded on each inference. true is recommended for best performance and is the default.

to_nhwc: Option<bool>tuning_type: Option<TuningType>tuning_file_path: Option<String>

Path to AutoTVM or Ansor tuning file which gives specifications for given model and target for the best performance.

input_names: Option<String>input_shapes: Option<String>
Implementations

impl TVM

pub fn build(self) -> ExecutionProviderDispatch