ort::memory

Struct Allocator

pub struct Allocator { /* private fields */ }

A device allocator used to manage the allocation of Values.

Direct allocation

Allocator can be used to directly allocate device memory. This can be useful if you have a postprocessing step that runs on the GPU.

let allocator = Allocator::new(
	&session,
	MemoryInfo::new(AllocationDevice::CUDA, 0, AllocatorType::Device, MemoryType::Default)?
)?;

let mut tensor = Tensor::<f32>::new(&allocator, [1_usize, 3, 224, 224])?;
// Here, `data_ptr` is a pointer to **device memory** inaccessible to the CPU; we'll need another crate, like
// `cudarc`, to access it.
let data_ptr = tensor.data_ptr_mut();

Note that ort does not facilitate the transfer of data between host & device outside of session inputs & outputs; you’ll need to use a separate crate for that, like cudarc for CUDA.

Pinned allocation

Memory allocated on the host CPU is often pageable and may reside on the disk (swap memory). Transferring pageable memory to another device is slow because the device has to go through the CPU to access the memory. Many execution providers thus provide a “pinned” allocator type, which allocates unpaged CPU memory that the device is able to access directly, bypassing the CPU and allowing for faster host-to-device data transfer.

If you create a session with a device allocator that supports pinned memory, like CUDA or ROCm, you can create an allocator for that session, and use it to allocate tensors with faster pinned memory:

let allocator = Allocator::new(
	&session,
	MemoryInfo::new(AllocationDevice::CUDA_PINNED, 0, AllocatorType::Device, MemoryType::CPUInput)?
)?;

// Create a tensor with our pinned allocator.
let mut tensor = Tensor::<f32>::new(&allocator, [1_usize, 3, 224, 224])?;
let data = tensor.extract_tensor_mut();
// ...fill `data` with data...

Implementations

impl Allocator

pub fn alloc<T>(&self, len: usize) -> Result<AllocatedBlock<'_>>

Allocates a block of memory, of size size_of::<T>() * len bytes, using this allocator. The memory will be automatically freed when the returned AllocatedBlock goes out of scope.

Returns an error if the allocation fails.

Example

let allocator = Allocator::default();
let mut mem = allocator.alloc::<i32>(5).unwrap();
unsafe {
	let ptr = mem.as_mut_ptr().cast::<i32>();
	*ptr.add(3) = 42;
};

pub unsafe fn free<T>(&self, ptr: *mut T)

Frees an object allocated by this allocator, given the object’s C pointer.

Safety

The pointer must have been allocated using this exact allocator’s alloc function.

This function is meant to be used in situations where the lifetime restrictions of AllocatedBlock are hard to work with.

let allocator = Allocator::default();
let mut mem = allocator.alloc::<i32>(5).unwrap();
unsafe {
	let ptr = mem.as_mut_ptr().cast::<i32>();
	*ptr.add(3) = 42;

	allocator.free(mem.into_raw());
};

pub fn memory_info<'a>(&'a self) -> &'a MemoryInfo<'a>

Returns the MemoryInfo describing this allocator.

pub fn new(session: &Session, memory_info: MemoryInfo<'_>) -> Result<Self>

Creates a new Allocator for the given session, to allocate memory on the device described in the MemoryInfo.