ort::operator

Trait Operator

pub trait Operator: Send + Sync {
    type Kernel<'attr>: Kernel + 'attr;

    const INPLACES: &[(u32, u32)] = _;
    const ALIASES: &[(u32, u32)] = _;

    // Required methods
    fn name(&self) -> &str;
    fn inputs(&self) -> impl IntoIterator<Item = OperatorInput>;
    fn outputs(&self) -> impl IntoIterator<Item = OperatorOutput>;
    fn create_kernel<'attr>(
        &self,
        kctx: &KernelContext<'attr>,
    ) -> Result<Self::Kernel<'attr>>;

    // Provided methods
    fn versions(&self) -> RangeInclusive<u32> { ... }
    fn execution_provider_type(&self) -> Option<&str> { ... }
    fn infer_shape(&self, ctx: &mut ShapeInferenceContext) -> Result<()> { ... }
}

A custom operator descriptor, which describes the expected inputs & outputs of a graph operator.

Operators are bound to OperatorDomains. Multiple operators can have the same name as long as they have different input/output types, in which case the exact operator will be picked depending on the input/output types. If you want to, for example, define a Sort operator that can accept either a single f32 or i64 tensor input, you’ll need to define 2 separate operators (which can be done via a macro); but both of these Operator structs can return the same name in Operator::name so that they are usable as simply my.domain:Sort in the graph.

Provided Associated Constants

const INPLACES: &[(u32, u32)] = _

A list of which inputs can be reused for an output if the shapes are identical; i.e. the operation for those outputs can be performed inplace, allowing the graph optimizer to save memory.

The tuple is (input_idx, output_idx).

A Mul operator with inputs A (tensor) and B (scalar) would set this to &[(0, 0)] since the A tensor can be reused for the output.

const ALIASES: &[(u32, u32)] = _

A list of which inputs are output without their data being changed, allowing for memory optimizations.

The tuple is (input_idx, output_idx).

The Reshape or Identity operator would set this to &[(0, 0)] because the input’s data doesn’t change (only the shape in the case of Reshape).

Required Associated Types

type Kernel<'attr>: Kernel + 'attr

Required Methods

fn name(&self) -> &str

Returns the name of the operator.

fn inputs(&self) -> impl IntoIterator<Item = OperatorInput>

fn outputs(&self) -> impl IntoIterator<Item = OperatorOutput>

fn create_kernel<'attr>( &self, kctx: &KernelContext<'attr>, ) -> Result<Self::Kernel<'attr>>

Provided Methods

fn versions(&self) -> RangeInclusive<u32>

Returns the domain version range supported by this operator.

fn execution_provider_type(&self) -> Option<&str>

Returns the internal name of the execution provider this operator runs on, e.g. "CUDAExecutionProvider" (see ExecutionProvider::name).

If the returned type is not None, and the execution provider used by the session matches this operator’s EP type, the value will not be copied to the CPU and you may use functions like Tensor::data_ptr to access the underlying device memory, and KernelContext::compute_stream to access the GPU compute stream.

fn infer_shape(&self, ctx: &mut ShapeInferenceContext) -> Result<()>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".