ort::session

Struct Session

pub struct Session { /* private fields */ }

An ONNX Runtime graph to be used for inference.

Use SessionBuilder/Session::builder to create a session, and Session::run to perform inference.

let mut session = Session::builder()?.commit_from_file("tests/data/upsample.onnx")?;
let input = ndarray::Array4::<f32>::zeros((1, 64, 64, 3));
let outputs = session.run(ort::inputs![TensorRef::from_array_view(&input)?])?;

Why does run take &mut self?

run_* methods take &mut self not because they mutate internal state, but because various ONNX Runtime internals – like certain EP allocators or statistics trackers – are not thread safe. Previous versions of ort that allowed concurrent inference often saw crashes and memory corruption, hence the change.

If you need concurrent inference, we recommend creating one session per thread or batching multiple requests at once until ONNX Runtime’s thread safety issues are fixed.

Implementations

impl Session

pub fn inputs(&self) -> &[Outlet]

pub fn outputs(&self) -> &[Outlet]

pub fn builder() -> Result<SessionBuilder>

Creates a new SessionBuilder.

pub fn allocator(&self) -> &Allocator

Returns this session’s Allocator.

pub fn create_binding(&self) -> Result<IoBinding>

Creates a new IoBinding for this session.

pub fn inner(&self) -> Arc<SharedSessionInner>

Get a shared (Arc’d) reference to the underlying SharedSessionInner, which holds the ort_sys::OrtSession pointer and the session allocator.

pub fn overridable_initializers(&self) -> Vec<OverridableInitializer>

Returns a list of initializers which are overridable (i.e. also graph inputs).

let session = Session::builder()?.commit_from_file("tests/data/overridable_initializer.onnx")?;

let mut overridable_initializers = session.overridable_initializers();
assert_eq!(overridable_initializers.len(), 1);
let f1 = overridable_initializers.pop().unwrap();
assert_eq!(f1.name(), "F1");
assert!(f1.dtype().is_tensor());

pub fn run<'s, 'i, 'v: 'i, const N: usize>( &'s mut self, input_values: impl Into<SessionInputs<'i, 'v, N>>, ) -> Result<SessionOutputs<'s>>

Run input data through the ONNX graph, performing inference.

See crate::inputs! for a convenient macro which will help you create your session inputs from ndarrays or other data. You can also provide a Vec, array, or HashMap of Values if you create your inputs dynamically.

let mut session = Session::builder()?.commit_from_file("tests/data/upsample.onnx")?;
let input = ndarray::Array4::<f32>::zeros((1, 64, 64, 3));
let outputs = session.run(ort::inputs![TensorRef::from_array_view(&input)?])?;

pub fn run_with_options<'r, 's: 'r, 'i, 'v: 'i, O: SelectedOutputMarker, const N: usize>( &'s mut self, input_values: impl Into<SessionInputs<'i, 'v, N>>, run_options: &'r RunOptions<O>, ) -> Result<SessionOutputs<'r>>

Run input data through the ONNX graph, performing inference, with a RunOptions struct. The most common usage of RunOptions is to allow the session run to be terminated from a different thread.

let run_options = Arc::new(RunOptions::new()?);

let run_options_ = Arc::clone(&run_options);
std::thread::spawn(move || {
	let _ = run_options_.terminate();
});

let res = session.run_with_options(ort::inputs![&input], &*run_options);
// upon termination, the session will return an `Error::SessionRun` error.`
assert_eq!(
	&res.unwrap_err().to_string(),
	"Failed to run inference on model: Exiting due to terminate flag being set to true."
);

pub fn run_binding<'b, 's: 'b>( &'s mut self, binding: &'b IoBinding, ) -> Result<SessionOutputs<'b>>

pub fn run_binding_with_options<'r, 'b, 's: 'b>( &'s mut self, binding: &'b IoBinding, run_options: &'r RunOptions<NoSelectedOutputs>, ) -> Result<SessionOutputs<'b>>

pub fn run_async<'r, 's: 'r, 'i, 'v: 'i + 's, O: SelectedOutputMarker, const N: usize>( &'s mut self, input_values: impl Into<SessionInputs<'i, 'v, N>>, run_options: &'r RunOptions<O>, ) -> Result<InferenceFut<'r, 'v>>

Asynchronously run input data through the ONNX graph, performing inference.

Inference will be performed on a thread in the session’s thread pool. Thus, the session must have been configured to have multiple intra-op threads; see SessionBuilder::with_intra_threads.

See crate::inputs! for a convenient macro which will help you create your session inputs from ndarrays or other data. You can also provide a Vec, array, or HashMap of Values if you create your inputs dynamically.

let mut session = Session::builder()?.with_intra_threads(2)?.commit_from_file("tests/data/upsample.onnx")?;
let input = ndarray::Array4::<f32>::zeros((1, 64, 64, 3));
let options = RunOptions::new()?;
let outputs = session.run_async(ort::inputs![TensorRef::from_array_view(&input)?], &options)?.await?;

pub fn metadata(&self) -> Result<ModelMetadata<'_>>

Gets the session model metadata. See ModelMetadata for more info.

pub fn profiling_start_ns(&self) -> Result<u64>

Returns the time that profiling was started, in nanoseconds.

pub fn end_profiling(&mut self) -> Result<String>

Ends profiling for this session. Returns the file name of the finalized profile.

Note that this must be explicitly called at the end of profiling, otherwise the profiling file will be empty.

pub fn set_workload_type(&mut self, workload_type: WorkloadType) -> Result<()>

Sets this session’s workload type to instruct execution providers to prioritize performance or efficiency.

let mut session = Session::builder()?.commit_from_file("tests/data/upsample.onnx")?;
session.set_workload_type(WorkloadType::Efficient)?;

let input = ndarray::Array4::<f32>::zeros((1, 64, 64, 3));
let outputs = session.run(ort::inputs![TensorRef::from_array_view(&input)?])?;

pub fn opset_for_domain(&self, domain: impl AsRef<str>) -> Option<u32>

Returns the version of the opset domain used by the model.

Requires the Model Editor API to be supported by the backend.

let session = Session::builder()?.commit_from_file("tests/data/lora_model.onnx")?;
assert_eq!(session.opset_for_domain(ort::editor::ONNX_DOMAIN), Some(21));