ort

Trait AsPointer

pub trait AsPointer {
    type Sys;

    // Required method
    fn ptr(&self) -> *const Self::Sys;

    // Provided method
    fn ptr_mut(&mut self) -> *mut Self::Sys { ... }
}

Trait to access raw pointers from safe types which wrap unsafe ort_sys types.

Required Associated Types

type Sys

This safe type’s corresponding ort_sys type.

Required Methods

fn ptr(&self) -> *const Self::Sys

Returns the underlying ort_sys type pointer this safe type wraps. The pointer is guaranteed to be non-null.

Provided Methods

fn ptr_mut(&mut self) -> *mut Self::Sys

Returns the underlying ort_sys type pointer this safe type wraps as a mutable pointer. The pointer is guaranteed to be non-null.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".