ort::util

Function preload_dylib

pub fn preload_dylib<P: AsFilename>(path: P) -> Result<(), Error>

Preloads the dynamic library at the given path.

Internally, some ONNX Runtime execution providers load DLLs with specific names (like onnxruntime_providers_cuda). There is no dedicated API to configure where these libraries are loaded from; the preload_dylib function offers a way to configure library paths without resorting to other tricks like rpath.

This is because most operating systems will keep a list of currently loaded dylibs in memory and fall back to an already loaded dylib if its file name matches what is requested. If we explicitly load libfoo.so from an absolute path, then any subsequent requests to load a library called foo will use the libfoo.so we already loaded, instead of searching the system for a foo library.

See also [crate::ep::cuda::preload_dylibs], a helper that uses preload_dylib to load all required dependencies of the CUDA execution provider.

use std::env;

// Use the DirectML DLL we ship alongside the application.
// This needs to come before we use any other `ort` API.
let application_dir = current_exe().unwrap();
let _ = ort::util::preload_dylib(application_dir.join("DirectML.dll"));