ort

Module value

Values are data containers used as inputs/outputs in ONNX Runtime graphs.

The most common type of value is Tensor:

// Create a tensor from a raw data vector
let tensor = Tensor::from_array(([1usize, 2, 3], vec![1.0_f32, 2.0, 3.0, 4.0, 5.0, 6.0].into_boxed_slice()))?;

// Create a tensor from an `ndarray::Array`
#[cfg(feature = "ndarray")]
let tensor = Tensor::from_array(ndarray::Array4::<f32>::zeros((1, 16, 16, 3)))?;

ONNX Runtime also supports Sequences and Maps, though they are less commonly used.

Structs

DynMapValueType

DynSequenceValueType

DynTensorValueType

DynValueTypeMarker

The dynamic type marker, used for values which can be of any type.

MapValueType

Outlet

The input/output to a session, node, or operator, consisting of a name and expected data type.

SequenceValueType

Shape

SymbolicDimensions

TensorArrayDataParts

TensorValueType

Value

A Value contains data for inputs/outputs in ONNX Runtime graphs. Values can be a Tensor, Sequence (aka array/vector), or Map.

ValueRef

A temporary version of a Value with a lifetime specifier.

ValueRefMut

A mutable temporary version of a Value with a lifetime specifier.

Enums

TensorElementType

Enum mapping ONNX Runtime’s supported tensor data types.

ValueType

The type of a Value, or a session input/output.

Traits

DefiniteTensorValueTypeMarker

Represents a type that is either a DynTensor or Tensor.

DowncastableTarget

Represents a type that a DynValue can be downcast to.

IntoShape

IntoTensorElementType

Trait used to map Rust types (for example f32) to ONNX tensor element data types (for example Float).

MapValueTypeMarker

OwnedTensorArrayData

PrimitiveTensorElementType

A superset of IntoTensorElementType that represents traits whose underlying memory is identical between Rust and C++ (i.e., every type except String).

SequenceValueTypeMarker

TensorArrayData

TensorArrayDataMut

TensorValueTypeMarker

ToShape

Utf8Data

Adapter for common Rust string types to ONNX strings.

ValueTypeMarker

Marker trait used to determine what operations can and cannot be performed on a Value of a given type.

Type Aliases

DynMap

DynMapRef

DynMapRefMut

DynSequence

DynSequenceRef

DynSequenceRefMut

DynTensor

A tensor Value whose data type is unknown.

DynTensorRef

A reference to a tensor Value whose data type is unknown.

DynTensorRefMut

A mutable reference to a tensor Value whose data type is unknown.

DynValue

A dynamic value, which could be a Tensor, Sequence, or Map.

Map

MapRef

MapRefMut

Sequence

SequenceRef

SequenceRefMut

Tensor

A strongly-typed tensor Value.

TensorRef

A reference to a strongly-typed tensor Value.

TensorRefMut

A mutable reference to a strongly-typed tensor Value.