ort::value

Enum ValueType

pub enum ValueType {
    Tensor {
        ty: TensorElementType,
        shape: Shape,
        dimension_symbols: SymbolicDimensions,
    },
    Sequence(Box<ValueType>),
    Map {
        key: TensorElementType,
        value: TensorElementType,
    },
    Optional(Box<ValueType>),
}

The type of a Value, or a session input/output.

// `ValueType`s can be obtained from session inputs/outputs:
let input = &session.inputs()[0];
assert_eq!(
	input.dtype(),
	&ValueType::Tensor {
		ty: TensorElementType::Float32,
		// Our model's input has 3 dynamic dimensions, represented by -1
		shape: Shape::new([-1, -1, -1, 3]),
		// Dynamic dimensions may also have names.
		dimension_symbols: SymbolicDimensions::new([
			"unk__31".to_string(),
			"unk__32".to_string(),
			"unk__33".to_string(),
			String::default()
		])
	}
);

// ...or by `Value`s created in Rust or output by a session.
let value = Tensor::from_array(([5usize], vec![1_i64, 2, 3, 4, 5].into_boxed_slice()))?;
assert_eq!(
	value.dtype(),
	&ValueType::Tensor {
		ty: TensorElementType::Int64,
		shape: Shape::new([5]),
		dimension_symbols: SymbolicDimensions::new([String::default()])
	}
);

Variants

Tensor

Value is a tensor/multi-dimensional array.

Fields

ty: TensorElementType

Element type of the tensor.

shape: Shape

Shape of the tensor. If an exact dimension is not known (i.e. a dynamic dimension as part of an Outlet), the dimension will be -1.

Actual tensor values (i.e. not Outlet definitions), which have a known dimension, will always have non-negative dimensions.

dimension_symbols: SymbolicDimensions

Sequence(Box<ValueType>)

A sequence (vector) of other Values.

Per ONNX spec, only sequences of tensors and maps are allowed.

Map

A map/dictionary from one element type to another.

Fields

key: TensorElementType

The map key type. Allowed types are:

TensorElementType::Int8

TensorElementType::Int16

TensorElementType::Int32

TensorElementType::Int64

TensorElementType::Uint8

TensorElementType::Uint16

TensorElementType::Uint32

TensorElementType::Uint64

TensorElementType::String

value: TensorElementType

The map value type.

Optional(Box<ValueType>)

An optional value, which may or may not contain a Value.

Implementations

impl ValueType

pub fn tensor_shape(&self) -> Option<&Shape>

Returns the shape of this value type if it is a tensor, or None if it is a sequence or map.

let value: DynValue = Tensor::from_array(([5usize], vec![1_i64, 2, 3, 4, 5].into_boxed_slice()))?.into_dyn();

let shape = value.dtype().tensor_shape().unwrap();
assert_eq!(**shape, [5]);

pub fn tensor_type(&self) -> Option<TensorElementType>

Returns the element type of this value type if it is a tensor, or None if it is a sequence or map.

let value = Tensor::from_array(([5usize], vec![1_i64, 2, 3, 4, 5].into_boxed_slice()))?;
assert_eq!(value.dtype().tensor_type(), Some(TensorElementType::Int64));

pub fn is_tensor(&self) -> bool

Returns true if this value type is a tensor.

pub fn is_sequence(&self) -> bool

Returns true if this value type is a sequence.

pub fn is_map(&self) -> bool

Returns true if this value type is a map.