ort::value

Struct ValueRef

pub struct ValueRef<'v, Type: ValueTypeMarker + ?Sized = DynValueTypeMarker> { /* private fields */ }

A temporary version of a Value with a lifetime specifier.

Implementations

impl<'a, T: PrimitiveTensorElementType + Debug> ValueRef<'a, TensorValueType<T>>

pub fn from_array_view( input: impl TensorArrayData<T> + 'a, ) -> Result<TensorRef<'a, T>>

Construct a tensor from borrowed data.

Borrowed tensors can be created from:

(with feature ndarray) a shared reference to a ndarray::CowArray (&CowArray<'_, T, D>) or ndarray::Array (&Array<T, D>);

(with feature ndarray) an ndarray::ArcArray or ndarray::ArrayView;

a tuple of (shape, data) where:

shape is one of Vec<I>, [I; N] or &[I], where I is i64 or usize, and

data is one of &[T], Arc<[T]>, or Arc<Box<[T]>>.

// Create a tensor from a raw data vector
let data = vec![1.0_f32, 2.0, 3.0, 4.0, 5.0, 6.0];
let tensor = TensorRef::from_array_view(([1usize, 2, 3], &*data))?;

// Create a tensor from an `ndarray::Array`
let array = ndarray::Array4::<f32>::zeros((1, 16, 16, 3));
let tensor = TensorRef::from_array_view(array.view())?;

When passing an ndarray type, the data must have a contiguous memory layout, or else an error will be returned. See ndarray::ArrayRef::as_standard_layout to convert an array to a contiguous layout.

impl<'v, Type: ValueTypeMarker + ?Sized> ValueRef<'v, Type>

pub fn downcast<OtherType: ValueTypeMarker + DowncastableTarget + ?Sized>( self, ) -> Result<ValueRef<'v, OtherType>>

Attempts to downcast a temporary dynamic value (like DynValue or DynTensor) to a more strongly typed variant, like TensorRef<T>.

pub fn try_upgrade(self) -> Result<Value<Type>, Self>

Attempts to upgrade this ValueRef to an owned Value holding the same data.

pub fn into_dyn(self) -> ValueRef<'v, DynValueTypeMarker>