ort::value

Struct ValueRefMut

pub struct ValueRefMut<'v, Type: ValueTypeMarker + ?Sized = DynValueTypeMarker> { /* private fields */ }

A mutable temporary version of a Value with a lifetime specifier.

Implementations

impl<'a, T: PrimitiveTensorElementType + Debug> ValueRefMut<'a, TensorValueType<T>>

pub fn from_array_view_mut( input: impl TensorArrayDataMut<T>, ) -> Result<TensorRefMut<'a, T>>

Construct a mutable tensor view from borrowed data. Modifying data through this view will modify the underlying buffer as well.

Mutably borrowed tensors can be created from:

(with feature ndarray) an exclusive reference to an ndarray::Array (&mut Array<T, D>);

(with feature ndarray) an ndarray::ArrayViewMut;

a tuple of (shape, &mut [T]), where shape is one of Vec<I>, [I; N] or &[I], where I is i64 or usize.

// Create a tensor from a raw data vector
let mut data = vec![1.0_f32, 2.0, 3.0, 4.0, 5.0, 6.0];
let tensor = TensorRefMut::from_array_view_mut(([1usize, 2, 3], &mut *data))?;

// Create a tensor from an `ndarray::Array`
let mut array = ndarray::Array4::<f32>::zeros((1, 16, 16, 3));
let tensor = TensorRefMut::from_array_view_mut(array.view_mut())?;

When passing an ndarray type, the data must have a contiguous memory layout, or else an error will be returned. See ndarray::ArrayRef::as_standard_layout to convert an array to a contiguous layout.

pub unsafe fn from_raw( info: MemoryInfo, data: *mut c_void, shape: Shape, ) -> Result<TensorRefMut<'a, T>>

Create a mutable tensor view from a raw pointer and shape.

The length of data is determined by T and the given shape, so the given buffer must be at least shape.num_elements() * size_of::<T>() bytes.

This function can be used to create data from raw device memory, e.g. to directly provide data to an execution provider. For instance, to create a tensor from a raw CUDA buffer using cudarc:

let device = CudaDevice::new(0)?;
let device_data = device.htod_sync_copy(&input_data)?;

let tensor: TensorRefMut<'_, f32> = unsafe {
	TensorRefMut::from_raw(
		MemoryInfo::new(AllocationDevice::CUDA, 0, AllocatorType::Device, MemoryType::Default)?,
		(*device_data.device_ptr() as usize as *mut ()).cast(),
		Shape::new([1, 3, 512, 512])
	)?
};

Safety

The pointer must be valid for the device description provided by MemoryInfo.

The returned tensor must outlive the data described by the data pointer.

impl<'v, Type: ValueTypeMarker + ?Sized> ValueRefMut<'v, Type>

pub fn downcast<OtherType: ValueTypeMarker + DowncastableTarget + ?Sized>( self, ) -> Result<ValueRefMut<'v, OtherType>>

Attempts to downcast a temporary mutable dynamic value (like DynValue or DynTensor) to a more strongly typed variant, like TensorRefMut<T>.

pub fn try_upgrade(self) -> Result<Value<Type>, Self>

Attempts to upgrade this ValueRefMut to an owned Value holding the same data.

pub fn into_dyn(self) -> ValueRefMut<'v, DynValueTypeMarker>