ort::value

Type Alias DynTensor

pub type DynTensor = Value<DynTensorValueType>;

A tensor Value whose data type is unknown.

Aliased Type

pub struct DynTensor { /* private fields */ }

Implementations

impl DynTensor

pub fn new( allocator: &Allocator, data_type: TensorElementType, shape: impl IntoShape, ) -> Result<DynTensor>

Construct a tensor via a given allocator with a given shape and datatype. The data in the tensor will be uninitialized.

This can be used to create a tensor with data on a certain device. For example, to create a tensor with pinned (CPU) memory for use with CUDA:

let allocator = Allocator::new(
	&session,
	MemoryInfo::new(AllocationDevice::CUDA_PINNED, 0, AllocatorType::Device, MemoryType::CPUInput)?
)?;

let mut img_input = DynTensor::new(&allocator, TensorElementType::Float32, [1_usize, 128, 128, 3])?;