ort::value

Type Alias DynValue

pub type DynValue = Value<DynValueTypeMarker>;

A dynamic value, which could be a Tensor, Sequence, or Map.

To attempt to convert a dynamic value to a strongly typed value, use DynValue::downcast. You can also attempt to extract data from dynamic values directly using try_extract_* methods; see Value for more information.

Aliased Type

pub struct DynValue { /* private fields */ }