ort::value

Type Alias Tensor

pub type Tensor<T> = Value<TensorValueType<T>>;

A strongly-typed tensor Value.

Aliased Type

pub struct Tensor<T> { /* private fields */ }

Implementations

impl Tensor<String>

pub fn from_string_array<T: Utf8Data>( input: impl TensorArrayData<T>, ) -> Result<Tensor<String>>

Construct a Tensor from an array of strings.

String tensors can be created from:

(with feature ndarray) a shared reference to a ndarray::CowArray (&CowArray<'_, T, D>) or ndarray::Array (&Array<T, D>);

(with feature ndarray) an ndarray::ArcArray or ndarray::ArrayView;

a tuple of (shape, data) where:

shape is one of Vec<I>, [I; N] or &[I], where I is i64 or usize, and

data is one of &[T], Arc<[T]>, or Arc<Box<[T]>>.

// Create a string tensor from a raw data vector
let data = vec!["hello", "world"];
let value = Tensor::from_string_array(([data.len()], &*data))?;

// Create a string tensor from an `ndarray::Array`
#[cfg(feature = "ndarray")]
let value = Tensor::from_string_array(&ndarray::Array::from_shape_vec((1,), vec!["document".to_owned()]).unwrap())?;

impl<T: PrimitiveTensorElementType + Debug> Tensor<T>

pub fn new(allocator: &Allocator, shape: impl IntoShape) -> Result<Tensor<T>>

Construct a tensor via a given allocator with a given shape and datatype. The data in the tensor will be uninitialized.

This can be used to create a tensor with data on a certain device. For example, to create a tensor with pinned (CPU) memory for use with CUDA:

let allocator = Allocator::new(
	&session,
	MemoryInfo::new(AllocationDevice::CUDA_PINNED, 0, AllocatorType::Device, MemoryType::CPUInput)?
)?;

let mut img_input = Tensor::<f32>::new(&allocator, [1_usize, 128, 128, 3])?;

pub fn from_array(input: impl OwnedTensorArrayData<T>) -> Result<Tensor<T>>

Construct an owned tensor from an array of data.

Owned tensors can be created from:

(with feature ndarray) an owned ndarray::Array, or

a tuple of (shape, data) where:

shape is one of Vec<I>, [I] or &[I], where I is i64 or usize, and

data is one of Vec<T> or Box<[T]>.

// Create a tensor from a raw data vector
let tensor = Tensor::from_array(([1usize, 2, 3], vec![1.0_f32, 2.0, 3.0, 4.0, 5.0, 6.0].into_boxed_slice()))?;

// Create a tensor from an `ndarray::Array`
#[cfg(feature = "ndarray")]
let tensor = Tensor::from_array(ndarray::Array4::<f32>::zeros((1, 16, 16, 3)))?;

When passing an ndarray::Array, the array may be copied in order to convert it to a contiguous layout if it is not already. When creating a tensor from a Vec or boxed slice, the data is assumed to already be in contiguous layout.

Creating string tensors requires a separate method; see Tensor::from_string_array.

impl<T: PrimitiveTensorElementType + Debug> Tensor<T>

pub fn extract_array(&self) -> ArrayViewD<'_, T>

Extracts the underlying data into a read-only ndarray::ArrayView.

let array = ndarray::Array4::<f32>::ones((1, 16, 16, 3));
let tensor = TensorRef::from_array_view(&array)?;

let extracted = tensor.extract_array();
assert_eq!(array.view().into_dyn(), extracted);

pub fn extract_array_mut(&mut self) -> ArrayViewMutD<'_, T>

Extracts the underlying data into a mutable ndarray::ArrayViewMut.

let mut array = ndarray::Array4::<f32>::ones((1, 16, 16, 3));
{
	let mut tensor = TensorRefMut::from_array_view_mut(array.view_mut())?;
	let mut extracted = tensor.extract_array_mut();
	extracted[[0, 0, 0, 1]] = 0.0;
}

assert_eq!(array[[0, 0, 0, 1]], 0.0);

pub fn extract_tensor(&self) -> (&Shape, &[T])

Extracts the underlying data into a view tuple, consisting of the tensor’s Shape and an immutable view into its data.

let array = vec![1_i64, 2, 3, 4, 5];
let tensor = TensorRef::from_array_view(([array.len()], &*array))?;

let (extracted_shape, extracted_data) = tensor.extract_tensor();
assert_eq!(extracted_data, &array);
assert_eq!(**extracted_shape, [5]);

pub fn extract_tensor_mut(&mut self) -> (&Shape, &mut [T])

Extracts the underlying data into a view tuple, consisting of the tensor’s shapes and a mutable view into its data.

let mut original_array = vec![1_i64, 2, 3, 4, 5];
{
	let mut tensor = TensorRefMut::from_array_view_mut(([original_array.len()], &mut *original_array))?;
	let (extracted_shape, extracted_data) = tensor.extract_tensor_mut();
	extracted_data[2] = 42;
}
assert_eq!(original_array, [1, 2, 42, 4, 5]);

impl<T: IntoTensorElementType + Debug> Tensor<T>

pub fn upcast(self) -> DynTensor

Converts from a strongly-typed Tensor<T> to a type-erased DynTensor.

let tensor = Tensor::<f32>::new(&Allocator::default(), [1_usize, 3, 224, 224])?;
let tensor_dyn = tensor.upcast();
assert!(tensor_dyn.try_extract_tensor::<f32>().is_ok());
assert!(tensor_dyn.try_extract_tensor::<i64>().is_err());

pub fn upcast_ref(&self) -> DynTensorRef<'_>

Creates a type-erased DynTensorRef from a strongly-typed Tensor<T>.

let tensor = Tensor::<f32>::new(&Allocator::default(), [1_usize, 3, 224, 224])?;
let tensor_dyn = tensor.upcast_ref();

let (_, original_extract) = tensor.extract_tensor();
let (_, ref_extract) = tensor_dyn.try_extract_tensor::<f32>()?;
assert_eq!(original_extract, ref_extract);

pub fn upcast_mut(&mut self) -> DynTensorRefMut<'_>

Converts from a strongly-typed Tensor<T> to a mutable reference to a type-erased DynTensor.

let mut tensor = Tensor::<i64>::from_array((vec![5], vec![1, 2, 3, 4, 5]))?;
let mut tensor_dyn = tensor.upcast_mut();

let (_, mut_view) = tensor_dyn.try_extract_tensor_mut::<i64>()?;
mut_view[3] = 0;

let (_, original_view) = tensor.extract_tensor();
assert_eq!(original_view, &[1, 2, 3, 0, 5]);