ort::value

Type Alias TensorRef

pub type TensorRef<'v, T> = ValueRef<'v, TensorValueType<T>>;

A reference to a strongly-typed tensor Value.

Aliased Type

pub struct TensorRef<'v, T> { /* private fields */ }

Implementations

impl<'a, T: PrimitiveTensorElementType + Debug> TensorRef<'a, T>

pub fn from_array_view( input: impl TensorArrayData<T> + 'a, ) -> Result<TensorRef<'a, T>>

Construct a tensor from borrowed data.

Borrowed tensors can be created from:

(with feature ndarray) a shared reference to a ndarray::CowArray (&CowArray<'_, T, D>) or ndarray::Array (&Array<T, D>);

(with feature ndarray) an ndarray::ArcArray or ndarray::ArrayView;

a tuple of (shape, data) where:

shape is one of Vec<I>, [I; N] or &[I], where I is i64 or usize, and

data is one of &[T], Arc<[T]>, or Arc<Box<[T]>>.

// Create a tensor from a raw data vector
let data = vec![1.0_f32, 2.0, 3.0, 4.0, 5.0, 6.0];
let tensor = TensorRef::from_array_view(([1usize, 2, 3], &*data))?;

// Create a tensor from an `ndarray::Array`
let array = ndarray::Array4::<f32>::zeros((1, 16, 16, 3));
let tensor = TensorRef::from_array_view(array.view())?;

When passing an ndarray type, the data must have a contiguous memory layout, or else an error will be returned. See ndarray::ArrayRef::as_standard_layout to convert an array to a contiguous layout.