ort

Module compiler

Provides ModelCompiler, which optimizes and creates EP graphs for an ONNX model ahead-of-time to greatly reduce startup time.

Many execution providers use a third-party graph-based neural network API (like CoreML, DirectML, CUDA graphs, TensorRT, etc.). The EPs must convert the ONNX graph to these device-native graphs, which can take a lot of work - especially when you consider that both ONNX Runtime and the EP backend are doing their own optimizations on their respective graphs! The end result is that sessions can often take a very long time to initialize - often multiple seconds and, in the case of more complex models, upwards of minutes. This is unacceptable for many applications, hence the need for ModelCompiler.

ModelCompiler is created like a normal Session, except instead of ever performing inference on the model, ModelCompiler only:

optimizes the ONNX model;

creates the EP graph(s);

saves EP states in a new model.

Loading a compiled model like one would with any other .onnx model (using the same EPs & options used to compile it) will greatly reduce the time required for the EP graph step. The model must be compiled on the end user’s system; you can’t distribute pre-compiled models.

let mut session_options = Session::builder()?.with_execution_providers([ep::CoreML::default()
	.with_model_format(ep::coreml::ModelFormat::MLProgram)
	.build()])?;

let compiled_path = PathBuf::from("model-coreml.compiled.onnx");
if !compiled_path.exists() {
	ModelCompiler::new(session_options.clone())?
		.with_model_from_file("model.onnx")?
		.compile_to_file(&compiled_path)?;
}

let session = session_options.commit_from_file(&compiled_path)?;

Structs

CompiledModel

ModelCompiler

Functions

compile_api

Returns a pointer to the global [ort_sys::OrtCompileApi] object, or errors if the Compile API is not supported by this backend.