ort

Module editor

Structs

Graph

A single graph in a Model, comprised of Nodes (and optional initializers, aka weights), and describing its inputs/outputs.

Model

Node

A single node in a Graph that performs a specific operation on its inputs.

Opset

Constants

ONNX_DOMAIN

The domain used for builtin ONNX operators under the canonical ai.onnx domain.

ONNX_ML_DOMAIN

The domain used for builtin ONNX ML operators under the canonical ai.onnx.ml domain.

Functions

editor_api

Returns a pointer to the global [ort_sys::OrtModelEditorApi] object, or errors if the Model Editor API is not supported by this backend.