ort::environment

Trait ThreadManager

pub trait ThreadManager {
    type Thread;

    // Required methods
    fn create(
        &self,
        work: impl FnOnce() + Send + 'static,
    ) -> Result<Self::Thread>;
    fn join(thread: Self::Thread) -> Result<()>;
}

Used for customizing the thread spawning process of a global thread pool or session thread pool. Could be used to add additional initialization/cleanup code to inference threads for better debugging/error handling.

Threads spawned by ThreadManager should be real threads, spawned directly via the operating system; they shouldn’t be spawned in another thread pool like rayon because sessions have their own (interfering) thread pool logic.

A very simple thread manager would be:

use std::thread::{self, JoinHandle};

use ort::environment::ThreadManager;

struct StdThreadManager;

impl ThreadManager for StdThreadManager {
	type Thread = JoinHandle<()>;

	fn create(&self, work: impl FnOnce() + Send + 'static) -> ort::Result<Self::Thread> {
		Ok(thread::spawn(move || {
			// ... maybe optional initialization code ...

			// threads must call work() to actually do the work the runtime needs
			work();

			// ... maybe optional destructor code ...
		}))
	}

	fn join(thread: Self::Thread) -> ort::Result<()> {
		let _ = thread.join();
		Ok(())
	}
}

Required Associated Types

type Thread

A handle to a spawned thread; used to join it later.

Required Methods

fn create(&self, work: impl FnOnce() + Send + 'static) -> Result<Self::Thread>

Spawns a thread.

The newly spawned thread must call work().

fn join(thread: Self::Thread) -> Result<()>

Wait for the thread to finish, like std::thread::JoinHandle::join.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.