ort

Module ep

ExecutionProviders provide hardware acceleration to Sessions.

Sessions can be configured with execution providers via SessionBuilder::with_execution_providers:

use ort::{ep, session::Session};

fn main() -> ort::Result<()> {
	let session = Session::builder()?
		.with_execution_providers([ep::CUDA::default().build()])?
		.commit_from_file("model.onnx")?;

	Ok(())
}

Re-exports

pub use self::cpu::CPU;

pub use self::cuda::CUDA;

pub use self::tensorrt::TensorRT;

pub use self::onednn::OneDNN;

pub use self::acl::ACL;

pub use self::openvino::OpenVINO;

pub use self::coreml::CoreML;

pub use self::rocm::ROCm;

pub use self::cann::CANN;

pub use self::directml::DirectML;

pub use self::tvm::TVM;

pub use self::nnapi::NNAPI;

pub use self::qnn::QNN;

pub use self::xnnpack::XNNPACK;

pub use self::armnn::ArmNN;Deprecated

pub use self::migraphx::MIGraphX;

pub use self::vitis::Vitis;

pub use self::rknpu::RKNPU;

pub use self::webgpu::WebGPU;

pub use self::azure::Azure;

pub use self::nvrtx::NVRTX;

Modules

acl

armnn

azure

cann

coreml

cpu

cuda

directml

migraphx

nnapi

nvrtx

onednn

openvino

qnn

rknpu

rocm

tensorrt

tvm

vitis

webgpu

xnnpack

Structs

ExecutionProviderDispatch

Dynamic execution provider container, used to provide a list of multiple types of execution providers when configuring execution providers for a SessionBuilder or EnvironmentBuilder.

ExecutionProviderLibrary

Handle to a loaded execution provider library, obtained from Environment::register_ep_library.

Enums

ArenaExtendStrategy

The strategy for extending the device memory arena.

RegisterError

Traits

ArbitrarilyConfigurableExecutionProvider

Trait used for execution providers that can have arbitrary configuration keys applied.

ExecutionProvider

Functions

get_gpu_device

Returns the ID of the GPU device being used by the active EP.

set_gpu_device

Sets the current GPU device of the active EP to the device specified by device_id.