ort::operator

Trait Operator

pub trait Operator: Send {
    // Required methods
    fn name(&self) -> &str;
    fn inputs(&self) -> Vec<OperatorInput>;
    fn outputs(&self) -> Vec<OperatorOutput>;
    fn create_kernel(
        &self,
        attributes: &KernelAttributes,
    ) -> Result<Box<dyn Kernel>>;

    // Provided methods
    fn execution_provider_type(&self) -> Option<&str> { ... }
    fn min_version(&self) -> i32 { ... }
    fn max_version(&self) -> i32 { ... }
    fn infer_shape(&self, ctx: &mut ShapeInferenceContext) -> Result<()> { ... }
}

A custom operator descriptor, which describes the expected inputs & outputs of a graph operator.

Operators are bound to OperatorDomains. Multiple operators can have the same name as long as they have different input/output types, in which case the exact operator will be picked depending on the input/output types. If you want to, for example, define a Sort operator that can accept either a single f32 or i64 tensor input, you’ll need to define 2 separate operators (which can be done via a macro); but both of these Operator structs can return the same name in Operator::name so that they are usable as simply my.domain:Sort in the graph.

Required Methods

fn name(&self) -> &str

Returns the name of the operator.

fn inputs(&self) -> Vec<OperatorInput>

fn outputs(&self) -> Vec<OperatorOutput>

fn create_kernel( &self, attributes: &KernelAttributes, ) -> Result<Box<dyn Kernel>>

Provided Methods

fn execution_provider_type(&self) -> Option<&str>

Returns the internal name of the execution provider this operator runs on, e.g. "CUDAExecutionProvider" (see ExecutionProvider::name).

If the returned type is not None, and the execution provider used by the session matches this operator’s EP type, the value will not be copied to the CPU and you may use functions like Tensor::data_ptr to access the underlying device memory, and KernelContext::compute_stream to access the GPU compute stream.

fn min_version(&self) -> i32

fn max_version(&self) -> i32

fn infer_shape(&self, ctx: &mut ShapeInferenceContext) -> Result<()>