ort

Module session

Contains Session, the main interface used to inference ONNX models.

let mut session = Session::builder()?.commit_from_file("tests/data/upsample.onnx")?;
let input = ndarray::Array4::<f32>::zeros((1, 64, 64, 3));
let outputs = session.run(ort::inputs![TensorRef::from_array_view(&input)?])?;

Re-exports

pub use self::input::SessionInputValue;

pub use self::input::SessionInputs;

pub use self::output::SessionOutputs;

pub use self::run_options::HasSelectedOutputs;

pub use self::run_options::NoSelectedOutputs;

pub use self::run_options::RunOptions;

pub use self::run_options::SelectedOutputMarker;

Modules

builder

input

metadata

output

run_options

Structs

InMemorySession

A Session where the graph data is stored in memory.

InferenceFut

OverridableInitializer

Session

An ONNX Runtime graph to be used for inference.

SharedSessionInner

Holds onto an [ort_sys::OrtSession] pointer and its associated allocator.

Enums

WorkloadType

Workload type, used to signal to execution providers whether to prioritize performance or efficiency.