ort

Module session

Contains Session, the main interface used to inference ONNX models.

let mut session = Session::builder()?.commit_from_file("tests/data/upsample.onnx")?;
let input = ndarray::Array4::<f32>::zeros((1, 64, 64, 3));
let outputs = session.run(ort::inputs![TensorRef::from_array_view(&input)?])?;

Modules

builder

Structs

Adapter

An input adapter, allowing for loading many static inputs from disk at once.

HasSelectedOutputs

Marks that a RunOptions was configured with an OutputSelector.

InMemorySession

A Session where the graph data is stored in memory.

InferenceFut

IoBinding

Enables binding of session inputs and/or outputs to pre-allocated memory.

ModelMetadata

Container for model metadata, including name & producer information.

NoSelectedOutputs

Marks that a RunOptions was not configured with an OutputSelector.

OutputSelector

Allows selecting/deselecting/preallocating the outputs of a Session inference call.

OverridableInitializer

RunOptions

Allows for finer control over session inference.

Session

An ONNX Runtime graph to be used for inference.

SessionOutputs

The outputs returned by a Session inference call.

SharedSessionInner

Holds onto an [ort_sys::OrtSession] pointer and its associated allocator.

Enums

SessionInputValue

SessionInputs

The inputs to a Session::run call.

WorkloadType

Workload type, used to signal to execution providers whether to prioritize performance or efficiency.

Traits

SelectedOutputMarker

Types that specify whether a RunOptions was configured with an OutputSelector.