ort::session::builder

Struct SessionBuilder

pub struct SessionBuilder { /* private fields */ }

Creates a session using the builder pattern.

Once configured, use the SessionBuilder::commit_from_file method to ‘commit’ the builder configuration into a Session.

let session = Session::builder()?
	.with_optimization_level(GraphOptimizationLevel::Level1)?
	.with_intra_threads(1)?
	.commit_from_file("tests/data/upsample.onnx")?;

Implementations

impl SessionBuilder

pub fn commit_from_file<P>(&mut self, model_path: P) -> Result<Session>
where P: AsRef<Path>,

Loads an ONNX model from a file and builds the session.

pub fn commit_from_memory_directly<'m>( &mut self, model_bytes: &'m [u8], ) -> Result<InMemorySession<'m>>

Load an ONNX graph from memory and commit the session For .ort models, we enable session.use_ort_model_bytes_directly. For more information, check Load ORT format model from an in-memory byte array.

If you wish to store the model bytes and the InMemorySession in the same struct, look for crates that facilitate creating self-referential structs, such as ouroboros.

pub fn commit_from_memory(&mut self, model_bytes: &[u8]) -> Result<Session>

Load an ONNX graph from memory and commit the session.

pub fn edit_from_file<P>( &mut self, model_filepath: P, ) -> Result<EditableSession<'_>>
where P: AsRef<Path>,

pub fn edit_from_memory( &mut self, model_bytes: &[u8], ) -> Result<EditableSession<'_>>

impl SessionBuilder

pub fn with_prepacking(self, enable: bool) -> BuilderResult

Enable/disable the usage of prepacking.

This option is enabled by default.

pub fn with_env_allocators(self) -> BuilderResult

Use allocators from the registered environment.

This option is disabled by default.

pub fn with_flush_to_zero(self) -> BuilderResult

Disables subnormal floats by enabling the denormals-are-zero and flush-to-zero flags for all threads in the session’s internal thread pool.

Subnormal floats are extremely small numbers very close to zero. Operations involving subnormal numbers can be very slow; enabling this flag will instead treat them as 0.0, giving faster & more consistent performance, but lower accuracy (in cases where subnormals are involved).

This option is disabled by default, as it may hurt model accuracy.

pub fn with_quant_qdq(self, enable: bool) -> BuilderResult

Enable/disable fusion for quantized models in QDQ (QuantizeLinear/DequantizeLinear) format.

This option is enabled by default for all EPs except DirectML.

pub fn with_double_qdq_remover(self, enable: bool) -> BuilderResult

Enable/disable the optimization step removing double QDQ nodes.

This option is enabled by default.

pub fn with_qdq_cleanup(self) -> BuilderResult

Enable the removal of Q/DQ node pairs once all QDQ handling has been completed.

This option is disabled by default.

pub fn with_approximate_gelu(self) -> BuilderResult

Enable fast tanh-based GELU approximation (like PyTorch’s nn.GELU(approximate='tanh')).

This option is disabled by default, as it may impact results.

pub fn with_cast_chain_elimination(self) -> BuilderResult

Enable the Cast chain elimination optimization.

This option is disabled by default, as it may impact results.

pub fn with_aot_inlining(self, enable: bool) -> BuilderResult

Enable/disable ahead-of-time function inlining.

This option is enabled by default.

pub fn with_disabled_optimizers( self, optimizers: impl AsRef<str>, ) -> BuilderResult

Accepts a comma-separated list of optimizers to disable.

pub fn with_device_allocated_initializers(self) -> BuilderResult

Enable using the device allocator for allocating initialized tensor memory, potentially bypassing arena allocators.

This option is disabled by default.

pub fn with_inter_op_spinning(self, enable: bool) -> BuilderResult

Enable/disable allowing the inter-op threads to spin for a short period before blocking.

This option is enabled by defualt.

pub fn with_intra_op_spinning(self, enable: bool) -> BuilderResult

Enable/disable allowing the intra-op threads to spin for a short period before blocking.

This option is enabled by defualt.

pub fn with_disable_cpu_fallback(self) -> BuilderResult

Disables falling back to the CPU for operations not supported by any other EP. Models with graphs that cannot be placed entirely on the EP(s) will fail to commit.

pub fn with_precise_qmm(self) -> BuilderResult

Uses slower U8U8 matrix multiplication in place of U8S8 matrix multiplication that could potentially overflow on x86-64 platforms without the VNNI extension.

This should only be enabled if you encounter overflow issues with quantized models.

pub fn with_dynamic_block_base(self, size: u32) -> BuilderResult

Enables dynamic thread block sizing with the given base block size.

impl SessionBuilder

pub fn with_execution_providers( self, execution_providers: impl AsRef<[ExecutionProviderDispatch]>, ) -> BuilderResult

Registers a list of execution providers for this session. Execution providers are registered in the order they are provided.

Execution providers will only work if the corresponding Cargo feature is enabled and ONNX Runtime was built with support for the corresponding execution provider. Execution providers that do not have their corresponding feature enabled will emit a warning.

Notes

Indiscriminate use of SessionBuilder::with_execution_providers in a library (e.g. always enabling CUDA) is discouraged unless you allow the user to configure the execution providers by providing a Vec of ExecutionProviderDispatches.

pub fn with_intra_threads(self, num_threads: usize) -> BuilderResult

Configure the session to use a number of threads to parallelize the execution within nodes. If ONNX Runtime was built with OpenMP (as is the case with Microsoft’s prebuilt binaries), this will have no effect on the number of threads used. Instead, you can configure the number of threads OpenMP uses via the OMP_NUM_THREADS environment variable.

For configuring the number of threads used when the session execution mode is set to Parallel, see SessionBuilder::with_inter_threads().

pub fn with_inter_threads(self, num_threads: usize) -> BuilderResult

Configure the session to use a number of threads to parallelize the execution of the graph. If nodes can be run in parallel, this sets the maximum number of threads to use to run them in parallel.

This has no effect when the session execution mode is set to Sequential.

For configuring the number of threads used to parallelize the execution within nodes, see SessionBuilder::with_intra_threads().

pub fn with_parallel_execution(self, parallel_execution: bool) -> BuilderResult

Enable/disable the parallel execution mode for this session. By default, this is disabled.

Parallel execution can improve performance for models with many branches, at the cost of higher memory usage. You can configure the amount of threads used to parallelize the execution of the graph via SessionBuilder::with_inter_threads().

pub fn with_optimization_level( self, opt_level: GraphOptimizationLevel, ) -> BuilderResult

Set the session’s optimization level. See GraphOptimizationLevel for more information on the different optimization levels.

pub fn with_optimized_model_path<S: AsRef<Path>>(self, path: S) -> BuilderResult

After performing optimization (configurable with SessionBuilder::with_optimization_level), serializes the newly optimized model to the given path (for ‘offline’ graph optimization).

Note that the file will only be created after the model is committed.

pub fn with_profiling<S: AsRef<Path>>(self, profiling_file: S) -> BuilderResult

Enables profiling. Profile information will be writen to profiling_file after profiling completes. See Session::end_profiling.

pub fn with_memory_pattern(self, enable: bool) -> BuilderResult

Enables/disables memory pattern optimization. Disable it if the input size varies, i.e., dynamic batch

pub fn with_allocator(self, info: MemoryInfo) -> BuilderResult

Configure this session to use a custom allocator, rather than the global default. This allocator is responsible for allocating the metadata associated with values – not the contents of the values themselves; that’s handled by the active execution providers. As such, only CPU-accessible allocators are allowed.

pub fn with_operator_library(self, lib_path: impl AsRef<Path>) -> BuilderResult

Registers a custom operator library at the given library path.

pub fn with_extensions(self) -> BuilderResult

Enables onnxruntime-extensions custom operators.

pub fn with_operators( self, domain: impl Into<Arc<OperatorDomain>>, ) -> BuilderResult

pub fn with_deterministic_compute(self, enable: bool) -> BuilderResult

Enables/disables deterministic computation.

The default (non-deterministic) kernels will typically use faster algorithms that may introduce slight variance. Enabling deterministic compute will output reproducible results, but may come at a performance penalty.

pub fn with_initializer( self, name: impl AsRef<str>, value: impl Into<Arc<DynValue>>, ) -> BuilderResult

pub fn with_external_initializer( self, name: impl AsRef<str>, value: impl Into<Arc<DynValue>>, ) -> BuilderResult

pub fn with_external_initializer_file_in_memory( self, file_name: impl AsRef<Path>, buffer: Cow<'static, [u8]>, ) -> BuilderResult

pub fn with_log_id(self, id: impl AsRef<str>) -> BuilderResult

pub fn with_dimension_override( self, name: impl AsRef<str>, size: i64, ) -> BuilderResult

pub fn with_dimension_override_by_denotation( self, denotation: impl AsRef<str>, size: i64, ) -> BuilderResult

pub fn with_prepacked_weights(self, weights: &PrepackedWeights) -> BuilderResult

pub fn with_independent_thread_pool(self) -> BuilderResult

Configures this environment to use its own thread pool instead of defaulting to the Environment’s global thread pool if one was defined.

pub fn with_no_environment_execution_providers(self) -> BuilderResult

pub fn with_thread_manager<T: ThreadManager + Any + 'static>( self, manager: T, ) -> BuilderResult

pub fn with_logger(self, logger: LoggerFunction) -> BuilderResult

Configures this session to use a custom logger function.

This will be called whenever a message pertaining to this session is to be logged, overriding the default log handler ([tracing] if the tracing feature is enabled, otherwise ONNX Runtime’s stdio logger).

use std::sync::Arc;

let mut session = Session::builder()?
	.with_logger(Arc::new(
		|level: ort::logging::LogLevel, category: &str, id: &str, code_location: &str, message: &str| {
			// ...
		}
	))?
	.commit_from_file("tests/data/upsample.onnx")?;

pub fn with_log_level(self, level: LogLevel) -> BuilderResult

Sets the severity level for messages logged by this session.

Note that when [tracing] integration is enabled via the tracing feature, the global log level takes precedence, i.e. if the application was initialized with ort’s log level set to warn via the RUST_LOG environment variable or similar, setting a session’s log severity level to verbose will still have it only log warn messages or higher.`

pub fn with_log_verbosity(self, verbosity: c_int) -> BuilderResult

Controls the level of verbosity for messages logged under LogLevel::Verbose; higher values = more verbose.

pub fn with_auto_device(self, policy: AutoDevicePolicy) -> BuilderResult

Automatically select & register an execution provider according to the given policy, based on available hardware devices.

For finer control over device selection, and to configure EP options, see SessionBuilder::with_devices.

let mut session = Session::builder()?
	// moar power!!1!
	.with_auto_device(AutoDevicePolicy::MaxPerformance)?
	.commit_from_file("tests/data/upsample.onnx")?;

pub fn with_devices<'e>( self, devices: impl IntoIterator<Item = Device<'e>>, options: Option<&[(String, String)]>, ) -> BuilderResult

Use a list of hardware devices automatically discovered by the environment via Environment::devices.

options can be specified to add EP options. Each EP option must be prefixed with the name of the EP (obtained by Device::ep) followed by ..

let env = Environment::current()?;

let options = vec![
	("CPUExecutionProvider.use_arena".to_string(), "1".to_string()),
	("XnnpackExecutionProvider.num_threads".to_string(), "4".to_string()),
];
let mut session = Session::builder()?
	.with_devices(env.devices().filter(|dev| dev.ty() == DeviceType::CPU), Some(&options))?
	.commit_from_file("tests/data/upsample.onnx")?;

impl SessionBuilder

pub fn new() -> Result<Self>

Creates a new session builder.

let session = Session::builder()?
	.with_optimization_level(GraphOptimizationLevel::Level1)?
	.with_intra_threads(1)?
	.commit_from_file("tests/data/upsample.onnx")?;

pub fn canceler(&self) -> LoadCanceler

Creates a signaler that can be used from another thread to cancel any in-progress commits.

let mut builder = Session::builder()?
	.with_optimization_level(GraphOptimizationLevel::Level1)?
	.with_intra_threads(1)?;

let canceler = builder.canceler();
thread::spawn(move || {
	thread::sleep(Duration::from_millis(500));
	// timeout if model hasn't loaded in 500ms
	let _ = canceler.cancel();
});

let session = builder.commit_from_file("tests/data/upsample.onnx")?;

pub fn with_config_entry( self, key: impl AsRef<str>, value: impl AsRef<str>, ) -> BuilderResult

Adds a custom configuration entry to the session.