ort::session

Struct IoBinding

pub struct IoBinding { /* private fields */ }

Enables binding of session inputs and/or outputs to pre-allocated memory.

IoBinding minimizes copies between a device (like a GPU) and the host (CPU) by allowing you to bind a certain input/output to a pre-allocated value on a specific device.

IoBinding is most suitable for:

An ensemble of models in which the output from one model is the input to another and does not need to pass through the CPU to perform additional processing.

Situations where an output should stay on a device (e.g. to perform additional hardware-accelerated processing).

Models that accept an input that does not change for multiple subsequent runs (like the conditional embedding for a diffusion model).

IoBinding will not provide any meaningful benefit for:

Models where every input changes with each invocation, such as a causal language model or object recognition model.

Pipelines that go straight from CPU -> GPU -> CPU.

Example

A diffusion model which takes a text condition input.

let mut text_encoder = Session::builder()?
	.with_execution_providers([ep::CUDA::default().build()])?
	.commit_from_file("text_encoder.onnx")?;
let mut unet = Session::builder()?
	.with_execution_providers([ep::CUDA::default().build()])?
	.commit_from_file("unet.onnx")?;

let text_condition = text_encoder
	.run(ort::inputs![Tensor::<i64>::from_array((
		vec![27],
		vec![
			23763, 15460, 473, 68, 312, 265, 17463, 4098, 304, 1077, 283, 198, 7676, 5976, 272, 285, 3609, 435,
			21680, 321, 265, 300, 1689, 64, 285, 4763, 64
		]
	))?])?
	.remove("output0")
	.unwrap();

let input_allocator = Allocator::new(
	&unet,
	MemoryInfo::new(AllocationDevice::CUDA_PINNED, 0, AllocatorType::Device, MemoryType::CPUInput)?
)?;
let mut latents = Tensor::<f32>::new(&input_allocator, [1_usize, 4, 64, 64])?;

let mut io_binding = unet.create_binding()?;
io_binding.bind_input("condition", &text_condition)?;

let output_allocator = Allocator::new(
	&unet,
	MemoryInfo::new(AllocationDevice::CUDA_PINNED, 0, AllocatorType::Device, MemoryType::CPUOutput)?
)?;
io_binding.bind_output("noise_pred", Tensor::<f32>::new(&output_allocator, [1_usize, 4, 64, 64])?)?;

for _ in 0..20 {
	io_binding.bind_input("latents", &latents)?;
	let noise_pred = unet.run_binding(&io_binding)?.remove("noise_pred").unwrap();

	let mut latents = latents.extract_array_mut();
	latents += &noise_pred.try_extract_array::<f32>()?;
}

IoBinding may provide a decent speedup in this example since the condition tensor is unchanging between runs. If we were to use normal session inference, the condition tensor would be needlessly copied with each invocation of unet.run(), and this copying can come with significant latency & overhead. With IoBinding, the condition tensor is only copied to the device once instead of 20 times.

Implementations

impl IoBinding

pub fn bind_input<T: ValueTypeMarker + ?Sized, S: Into<String>>( &mut self, name: S, ort_value: &Value<T>, ) -> Result<()>

Bind a Value to a session input.

Upon invocation, the value’s data will be queued to be copied to the device the session is allocated on. The copied data will be used as an input (specified by name) in all future invocations of Session::run_binding until the input is overridden (by calling IoBinding::bind_input again) or until all inputs are cleared (via IoBinding::clear_inputs or IoBinding::clear).

The data is only copied upon invocation of this function. Any changes to the given value afterwards will not affect the data seen by the session until the value is re-bound.

pub fn bind_output<T: ValueTypeMarker + ?Sized, S: Into<String>>( &mut self, name: S, ort_value: Value<T>, ) -> Result<()>

Bind a session output to a pre-allocated Value.

This allows for the pre-allocation and reuse of memory in the session output (see Tensor::new). Any subsequent runs via Session::run_binding will reuse the same tensor to store the output instead of creating a new one each time.

The output will be accessible in the value returned by Session::run_binding, under the name specified by name.

pub fn bind_output_to_device<S: Into<String>>( &mut self, name: S, mem_info: &MemoryInfo, ) -> Result<()>

Bind a session output to a device which is specified by mem_info. Used when the shape of the output is unknown.

Note that the same output buffer will be reused across runs, so the output must be .clone()d to avoid race conditions.

If the shape of the output is known, prefer IoBinding::bind_output instead.

pub fn clear_inputs(&mut self)

Clears all bound inputs specified by IoBinding::bind_input.

pub fn clear_outputs(&mut self)

Clears all bound outputs specified by IoBinding::bind_output or IoBinding::bind_output_to_device.

pub fn clear(&mut self)

Clears both the bound inputs & outputs; equivalent to IoBinding::clear_inputs followed by IoBinding::clear_outputs.

pub fn synchronize_inputs(&self) -> Result<()>

Synchronize all bound inputs, ensuring any pending asynchronous transfers are completed.

pub fn synchronize_outputs(&self) -> Result<()>

Synchronize all bound outputs, ensuring any pending asynchronous transfers are completed.

pub fn synchronize(&self) -> Result<()>

Synchronizes both inputs & outputs; equivalent to IoBinding::synchronize_inputs followed by IoBinding::synchronize_outputs.