ort::session

Struct RunOptions

pub struct RunOptions<O: SelectedOutputMarker = NoSelectedOutputs> { /* private fields */ }

Allows for finer control over session inference.

RunOptions provides three main features:

Run tagging: Each individual session run can have a uniquely identifiable tag attached with RunOptions::set_tag, which will show up in logs. This can be especially useful for debugging performance/errors in inference servers.

Termination: Allows for terminating an inference call from another thread; when RunOptions::terminate is called, any sessions currently running under that RunOptions instance will halt graph execution as soon as the termination signal is received. This allows for Session::run_async’s cancel-safety.

Output specification: Certain session outputs can be disabled or pre-allocated. Disabling an output might mean ONNX Runtime will not execute parts of the graph that are only used by that output. Pre-allocation can reduce expensive re-allocations by allowing you to use the same memory across runs.

RunOptions can be passed to most places where a session can be inferred, e.g. Session::run_with_options, Session::run_async, Session::run_binding_with_options. Some of these patterns (notably IoBinding) do not accept OutputSelector, hence RunOptions contains an additional type parameter that marks whether or not outputs have been selected.

Implementations

impl RunOptions

pub fn new() -> Result<RunOptions<NoSelectedOutputs>>

Creates a new RunOptions struct.

impl<O: SelectedOutputMarker> RunOptions<O>

pub fn with_outputs( self, outputs: OutputSelector, ) -> RunOptions<HasSelectedOutputs>

Select/deselect/preallocate outputs for this run.

See OutputSelector for more details.

let mut session = Session::builder()?.commit_from_file("tests/data/upsample.onnx")?;
let input = Tensor::<f32>::new(&Allocator::default(), [1_usize, 64, 64, 3])?;

let output0 = session.outputs()[0].name();
let options = RunOptions::new()?.with_outputs(
	// Disable all outputs...
	OutputSelector::no_default()
		// except for the first one...
		.with(output0)
		// and since this is a 2x upsampler model, pre-allocate the output to be twice as large.
		.preallocate(output0, Tensor::<f32>::new(&Allocator::default(), [1_usize, 128, 128, 3])?)
);

// `outputs[0]` will be the tensor we just pre-allocated.
let outputs = session.run_with_options(ort::inputs![input], &options)?;

pub fn with_tag(self, tag: impl AsRef<str>) -> Result<Self>

Sets a tag to identify this run in logs.

pub fn set_tag(&mut self, tag: impl AsRef<str>) -> Result<()>

Sets a tag to identify this run in logs.

pub fn tag(&self) -> Result<&str>

pub fn terminate(&self) -> Result<()>

Sets the termination flag for the runs associated with this RunOptions.

This function returns immediately (it does not wait for the session run to terminate). The run will terminate as soon as it is able to.

let run_options = Arc::new(RunOptions::new()?);

let run_options_ = Arc::clone(&run_options);
std::thread::spawn(move || {
	let _ = run_options_.terminate();
});

let res = session.run_with_options(ort::inputs![input], &*run_options);
// upon termination, the session will return an `Error::SessionRun` error.`
assert_eq!(
	&res.unwrap_err().to_string(),
	"Failed to run inference on model: Exiting due to terminate flag being set to true."
);

pub fn unterminate(&self) -> Result<()>

Resets the termination flag for the runs associated with RunOptions.

let run_options = Arc::new(RunOptions::new()?);

let run_options_ = Arc::clone(&run_options);
std::thread::spawn(move || {
	let _ = run_options_.terminate();
	// ...oops, didn't mean to do that
	let _ = run_options_.unterminate();
});

let res = session.run_with_options(ort::inputs![input], &*run_options);
assert!(res.is_ok());

pub fn add_config_entry( &mut self, key: impl AsRef<str>, value: impl AsRef<str>, ) -> Result<()>

Adds a custom configuration option to the RunOptions.

This can be used to, for example, configure the graph ID when using compute graphs with an execution provider like CUDA:

let mut run_options = RunOptions::new()?;
run_options.add_config_entry("gpu_graph_id", "1")?;

pub fn add_adapter(&mut self, adapter: &Adapter) -> Result<()>

pub fn set_log_level(&mut self, level: LogLevel) -> Result<()>

pub fn log_level(&self) -> Result<LogLevel>

pub fn set_log_verbosity(&mut self, verbosity: c_int) -> Result<()>

pub fn log_verbosity(&self) -> Result<i32>

pub fn disable_device_sync(&mut self) -> Result<()>