ort::value

Trait ValueTypeMarker

pub trait ValueTypeMarker { }

Marker trait used to determine what operations can and cannot be performed on a Value of a given type.

For example, Tensor::try_extract_tensor can only be used on Values with the TensorValueTypeMarker (which inherits this trait), i.e. Tensors, DynTensors, and DynValues.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.